Vector: reimplement without an additional abstraction of Chunks
Since implementing destruction of each of the `Clone` objects in
destructor of Vector, we need to only destruct variables within
`self.len`, not within `self.cap`. However, an intrinsic logic of `Chunks`
is aware of whether any of the values are initialized or not.

So, as a logical consequence, a destruction of array of `T` should be
a matter of `Vector``, not of `Chunks`.

To somehow abstract away pointer operations semantics, declarative
`array_*` functions were introduced.